Food_generation_Apllication/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
